[bfunet] convunext denoiser: decode-once pipeline, ppi 8->4, opt-in mixed precision
train_convunext_denoiser.py:
- create_dataset now decodes each image ONCE and crops patches_per_image patches
  from it (split load_and_preprocess_image into decode_full_image + random_crop_patch).
  Removes the per-patch re-read/re-decode of the same file (~8x fewer JPEG decodes/epoch);
  datasets are on a spinning HDD so this cuts real CPU + I/O load. deterministic=False +
  AUTOTUNE on the decode map to overlap read latency.
- patches_per_image default 8 -> 4 (dataclass + CLI); steps_per_epoch halves accordingly.
- Add --mixed-precision (mixed_float16) flag: global policy + fp32 output cast +
  LossScaleOptimizer + jit_compile=False. OFF by default. Benchmarked SLOWER than the
  fp32 default on base@256/b4 (RTX 4090, ~22 vs 36 img/s): the decoder's bilinear-upsample
  grad emits fp32, forcing XLA off, and for this conv-heavy U-Net losing XLA fusion
  outweighs the fp16 gain. Kept as a documented option (help text + docstring note).

layer_scale.py:
- LearnableMultiplier.call casts gamma to the input compute dtype, making it
  mixed_float16-safe (fp32-weight x fp16-activation was rejected by XLA in the gradient).
  No-op in fp32; existing fp32 unit tests unaffected.

+++ src/dl_techniques/layers/layer_scale.py
@@ -212,8 +212,12 @@ class LearnableMultiplier(keras.layers.Layer):
         :return: Scaled tensor with same shape as input.
         :rtype: keras.KerasTensor
         """
-        # Element-wise multiplication using Keras ops for backend compatibility
-        return ops.multiply(inputs, self.gamma)
+        # Element-wise multiplication using Keras ops for backend compatibility.
+        # Cast gamma (stored fp32, including under a mixed_float16 policy) to the input's
+        # compute dtype so fp16 activations multiply an fp16 scale. Without the cast the
+        # fp32-weight x fp16-activation mismatch is rejected by XLA in the gradient
+        # ("mixed precision disallowed"). No-op when dtypes already match (fp32 path).
+        return ops.multiply(inputs, ops.cast(self.gamma, inputs.dtype))
 
     def compute_output_shape(self, input_shape: Tuple[Optional[int], ...]) -> Tuple[Optional[int], ...]:
         """Compute the output shape of the layer.
